refactor: unify IdDocument usage across controllers and components fix: minor things
