+++ common/formatter.ts
@@ -1,5 +1,5 @@
 import { baseCurrencyMoneyToMoney, isValidDate } from './scripts.js'
-import { BaseCurrencyMoney, Locale, Money, baseCurrency } from './types.js'
+import { BaseCurrencyMoney, Locale, Money, baseCurrency, idDocumentToId } from './types.js'
 
 type MoneyStringOptions = { locale?: Locale; useExchangeRate?: boolean; func?: (x: number) => number; warning?: boolean }
 
@@ -110,7 +110,7 @@ class Formatter {
     this.setLocale(opts.locale)
     let amount = 0
     const money = baseCurrencyMoneyToMoney(baseMoney)
-    let currency = typeof money.currency === 'string' ? money.currency : money.currency._id
+    let currency = idDocumentToId(money.currency)
     if (opts.useExchangeRate && money.exchangeRate) {
       amount = money.exchangeRate.amount
       currency = baseCurrency._id
