+++ frontend/src/components/healthCareCost/HealthCareCostPage.vue
@@ -28,7 +28,7 @@
           :healthCareCost="modalObject"
           :loading="modalFormIsLoading"
           :owner="healthCareCost.owner._id"
-          :update-user-org="endpointPrefix !== '/examine'"
+          :update-user-org="endpointPrefix !== 'examine/'"
           :endpoint-prefix="endpointPrefix"
           @cancel="resetAndHide()"
           @add="editHealthCareCostDetails">
