+++ frontend/src/components/healthCareCost/forms/HealthCareCostForm.vue
@@ -91,6 +91,7 @@ export default defineComponent({
       return {
         name: '',
         patientName: '',
+        advances: [],
         owner: this.owner
       }
     },
